requirements.txt
alembic==1.6.5
astroid==2.6.2
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
boto3==1.14.34
botocore==1.17.34
click==7.1.2
colorama==0.4.4
docopt==0.6.2
docutils==0.15.2
flake8==3.9.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
iniconfig==1.1.1
isort==5.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.25
natsort==6.2.1
packaging==20.9
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.3
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
PyYAML==5.4
rope==0.19.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.4.18
toml==0.10.2
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
